pyproject.toml
[tool.poetry]
name = "langchain-crash-course"
version = "0.1.0"
description = "Everything you need to know to get started with LangChain"
authors = ["bhancock_ai <brandon@brandonhancock.io>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.12.2"

# LangChain Core & Ecosystem (updated for compatibility)
langchain-core = "^0.3.69"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-ollama = "^0.3.5"
langchain-google-genai = "^2.1.8"
langchain-groq = "^0.3.6"


# Other Dependencies
python-dotenv = "^1.0.1"
firestore = "^0.0.8"
tiktoken = "^0.7.0"
sentence-transformers = "^3.0.0"
bs4 = "^0.0.2"
firecrawl-py = "^0.0.13"
langchainhub = "^0.1.18"
wikipedia = "^1.4.0"
tavily-python = "^0.3.3"
langchain-huggingface = "^0.3.0"
langchain-chroma = "^0.2.4"
langchain-text-splitters = "^0.3.8"
langgraph = "^0.5.3"
langchain-openai = "^0.3.28"
pymupdf = "^1.26.3"

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"